feat: refactor type listener, move transaction types to a separate folder
